Separate visualization tests to avoid race condition
